Added casts to size calculations for long vectors.
git-svn-id: https://svn.r-project.org/R/trunk@60278 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/lapack/Lapack.c
@@ -80,9 +80,9 @@ static SEXP modLa_svd(SEXP jobu, SEXP jobv, SEXP x, SEXP s, SEXP u, SEXP v, SEXP
     xdims = INTEGER(coerceVector(getAttrib(x, R_DimSymbol), INTSXP));
     n = xdims[0];
     p = xdims[1];
-    xvals = (double *)R_alloc(n * p, sizeof(double));
+    xvals = (double *)R_alloc(n * (size_t)p, sizeof(double));
     /* work on a copy of x */
-    Memcpy(xvals, REAL(x), n * p);
+    Memcpy(xvals, REAL(x), n * (size_t)p);
 
     {
         int ldu = INTEGER(getAttrib(u, R_DimSymbol))[0], ldvt = INTEGER(getAttrib(v, R_DimSymbol))[0];
